+++ database/migrations/2023_10_19_110428_create_penjualan_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('penjualan', function (Blueprint $table) {
             $table->increments('id_penjualan');
             $table->string('tanggal',50);
-            $table->bigInteger('id_customer');
+            $table->unsignedInteger('id_customer');
             $table->integer('total_penjualan');
             $table->string('metode_pembayaran');
             $table->string('status_pembayaran');
