+++ database/migrations/2023_11_01_060417_create_detail_pembelians_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('detail_pembelians', function (Blueprint $table) {
             $table->increments('id_detail_pembelian');
-            $table->bigInteger('id_pembelian');
-            $table->bigInteger('id_barang');
+            $table->unsignedInteger('id_pembelian');
+            $table->unsignedInteger('id_barang');
             $table->integer('kuantitas');
             $table->integer('harga_beli');
             $table->timestamps();
